requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
langchain
langchain-core
langchain-google-genai
google-generativeai
tavily-python
pydantic
